feat: use html attributes instead of css class for defining touch (dnb-is-touch) + rename src/shared/tools to src/shared/global-heloers
