+++ packages/dnb-design-system-portal/gatsby-browser.js
@@ -3,7 +3,7 @@
  *
  */
 
-import { applyPageFocus } from 'dnb-ui-lib/src/shared/tools'
+import { applyPageFocus } from 'dnb-ui-lib/src/shared/global-helpers'
 
 // Load dev styles (to use hot reloading, we do have to import the styles in here)
 if (process.env.NODE_ENV === 'development') {
