+++ packages/dnb-design-system-portal/src/shared/parts/Layout.js
@@ -20,7 +20,7 @@ import ToggleGrid from '../menu/ToggleGrid'
 import {
   setPageFocusElement,
   scrollToLocationHashId
-} from 'dnb-ui-lib/src/shared/tools'
+} from 'dnb-ui-lib/src/shared/global-helpers'
 import { Logo } from 'dnb-ui-lib/src'
 
 class Layout extends PureComponent {
