+++ packages/dnb-ui-lib/src/patterns/main-nav/style/_main-nav.scss
@@ -20,7 +20,7 @@
 @include allAbove(large) {
   //Nav bar becomes fixed on large screens
   //so .main needs to push content a bit down
-  .no-touch .main {
+  :not([dnb-is-touch]) .main {
     padding-top: 3rem;
   }
 }
